feat(posts): Move index page into the folder to co-locate its components
