+++ app/routes/_index/components/PostsList.tsx
@@ -1,0 +1,3 @@
+import type {FC} from "react"
+
+export const PostsList: FC = () => {}
