+++ app/routes/_index/contexts/PostsContext.tsx
@@ -1,0 +1,20 @@
+import {createContext, useContext} from "react"
+
+import type {MaybeNull} from "../../../lib/types/MaybeNull.js"
+
+import {loader} from "../route.jsx"
+
+export type PostsContextData =
+  Awaited<ReturnType<Awaited<ReturnType<typeof loader>>["json"]>>
+
+export const PostsContext = createContext<MaybeNull<PostsContextData>>(null)
+
+export function usePostsContext(): PostsContextData {
+  const context = useContext(PostsContext)
+
+  if (!context) {
+    throw new Error("Unable to find PostsContext")
+  }
+
+  return context
+}
