+++ app/routes/_index/route.tsx
@@ -1,8 +1,10 @@
 import {json, type MetaFunction} from "@remix-run/node"
 import {useLoaderData} from "@remix-run/react"
 
-import {Post} from "../server/db/entities.js"
-import {withOrm} from "../server/lib/db/orm.js"
+import {Post} from "../../server/db/entities.js"
+import {withOrm} from "../../server/lib/db/orm.js"
+
+import {NoPosts} from "./components/NoPosts.jsx"
 
 export const loader = withOrm(async orm => {
   const [list, count] = await orm.em.findAndCount(Post, {}, {
@@ -25,10 +27,12 @@ const HomePage = () => {
   const {count} = useLoaderData<typeof loader>()
 
   return (
-    <div className="w-full p-5 laptop:w-laptop laptop:mx-auto">
-      <div className="text-center">
-        {count > 0 ? "Posts will apper here" : "No posts yet"}
-      </div>
+    <div className="flex flex-col w-full p-5 laptop:w-laptop laptop:mx-auto">
+      {
+        count > 0
+          ? <div>Posts will apper here</div>
+          : <NoPosts />
+      }
     </div>
   )
 }
