+++ app/routes/admin._index.tsx
@@ -1,5 +1,5 @@
 // TODO: Replace loader on this page with it's own when I add filters
-export {loader} from "./_index.jsx"
+export {loader} from "./_index/route.jsx"
 
 const AdminPage = () => <div>Admin dashboard will be here</div>
